Added methods for getting time and humidity from Smhi.

+++ src/main/java/se/iths/group/forecastserviceteamproject/smhi/Parameter.java
@@ -28,7 +28,7 @@ public class Parameter {
     @JsonProperty("unit")
     private String unit;
     @JsonProperty("values")
-    private List<Integer> values = null;
+    private List<Double> values = null;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
@@ -73,12 +73,12 @@ public class Parameter {
     }
 
     @JsonProperty("values")
-    public List<Integer> getValues() {
+    public List<Double> getValues() {
         return values;
     }
 
     @JsonProperty("values")
-    public void setValues(List<Integer> values) {
+    public void setValues(List<Double> values) {
         this.values = values;
     }
